feat: split stores into petition validation and progress
